utilize ts namespace
