+++ src/conn.ts
@@ -1,0 +1,9 @@
+import { getMessenger } from './sdk/messenger';
+
+/**
+ * The Messenger instance that sends/receives messages between game wrapper.
+ * @hidden
+ */
+const conn = getMessenger();
+
+export default conn;
